requirements.txt
google-generativeai 
gradio